added card service and testing

+++ src/app/shared/services/http.service.ts
@@ -18,25 +18,49 @@ export class HttpService {
     })
   };
 
-  getAll(url: string){
-    return this.http.get(url, this.httpOptions);
+  httpOptionsUnsecured = {
+    headers: new HttpHeaders({
+      'Content-Type':  'application/json',
+      'Accept': 'application/json'
+      //'Authorization': this.token
+    })
+  };
+
+  getAll(url: string, security: boolean){
+    if(security){
+      return this.http.get(url, this.httpOptions);
+    }
+    else{
+      return this.http.get(url, this.httpOptionsUnsecured);
+    }
   }
 
-  getById(url: string){
-    return this.http.get(url, this.httpOptions);
+  getById(url: string, security: boolean){
+    if(security){
+      return this.http.get(url, this.httpOptions);
+    }
+    return this.http.get(url, this.httpOptionsUnsecured);
   }
 
-  deleteById(url: string){
-    return this.http.delete(url, this.httpOptions);
+  deleteById(url: string, security: boolean){
+    if(security){
+      return this.http.delete(url, this.httpOptions);
+    }
+    return this.http.delete(url, this.httpOptionsUnsecured);
   }
 
-  create(url: string, obj: any){
-
-    return this.http.post(url, obj, this.httpOptions);
+  create(url: string, obj: any, security: boolean){
+    if(security){
+      return this.http.post(url, obj, this.httpOptions);
+    }
+    return this.http.post(url, obj, this.httpOptionsUnsecured);
   }
 
-  update(url: string, obj: any){
-    return this.http.put(url, obj, this.httpOptions);
+  update(url: string, obj: any, security: boolean){
+    if(security){
+      return this.http.put(url, obj, this.httpOptions);
+    }
+    return this.http.put(url, obj, this.httpOptionsUnsecured);
   }
 
 }
